syntax: Implement Mark API for GreenTreeBuilder
Instead of storing teh children as Firethorn did we can instead get away
with an array of marks on each node. Expired marks are cleared when
a mark is sliced.

The mark slice method is intended to allow slicing off a mark into a new
node to aid in the building of _closing_ tags. For closing tags we don't
always know which node they should be parented to. In this case we might
have to slice them off, unwind the stack, and then re-apply.